Исправить https://github.com/Eugenpak/java-kanban/pull/1#discussion_r1863328425, https://github.com/Eugenpak/java-kanban/pull/1#discussion_r1863322730, https://github.com/Eugenpak/java-kanban/pull/1#discussion_r1863317965
