+++ test/controllers/InMemoryHistoryManagerTest.java
@@ -173,56 +173,7 @@ class InMemoryHistoryManagerTest {
         assertNotEquals(2,listHistory.get(4).getId(), "Список не пустой");
     }
 
-    @Test
-    void removeEpicFromGetHistory() {
-        InMemoryHistoryManager inMemoryHistoryManager = new InMemoryHistoryManager();
-        assertEquals(0,inMemoryHistoryManager.getHistory().size(), "Список не пустой");
-        Epic epic1 = new Epic("E0","DE0",0,Status.NEW);
-        inMemoryHistoryManager.add(epic1);
-        ArrayList<Subtask> arraySubtask = new ArrayList<>();
-        for (int i = 1; i <= 5; i++) {
-            Subtask subtask = new Subtask("S" + i,"DS" + i,i,Status.NEW,0);
-            inMemoryHistoryManager.add(subtask);
-            arraySubtask.add(subtask);
-        }
-        epic1.setArraySubtask(arraySubtask);
-        inMemoryHistoryManager.add(epic1);
-        epic1 = new Epic("E6","DE6",6,Status.NEW);
-        inMemoryHistoryManager.add(epic1);
 
-        inMemoryHistoryManager.remove(0);
-        List<Task> listHistory = inMemoryHistoryManager.getHistory();
-        assertEquals(1,listHistory.size(), "Список не пустой");
-        assertEquals(6,listHistory.get(0).getId(), "Список не пустой");
-    }
-
-    @Test
-    void removeSubtaskFromGetHistory() {
-        InMemoryHistoryManager inMemoryHistoryManager = new InMemoryHistoryManager();
-        assertEquals(0,inMemoryHistoryManager.getHistory().size(), "Список не пустой");
-        Epic epic1 = new Epic("E0","DE0",0,Status.NEW);
-        inMemoryHistoryManager.add(epic1);
-        ArrayList<Subtask> arraySubtask = new ArrayList<>();
-        for (int i = 1; i <= 5; i++) {
-            Subtask subtask = new Subtask("S" + i,"DS" + i,i,Status.NEW,0);
-            inMemoryHistoryManager.add(subtask);
-            arraySubtask.add(subtask);
-        }
-        epic1.setArraySubtask(arraySubtask);
-        inMemoryHistoryManager.add(epic1);
-        epic1 = new Epic("E6","DE6",6,Status.NEW);
-        inMemoryHistoryManager.add(epic1);
-        List<Task> listHistory = inMemoryHistoryManager.getHistory();
-        Subtask subtaskAct = (Subtask) listHistory.get(0);
-        assertEquals(7,listHistory.size(), "Список не пустой");
-        assertEquals(1,subtaskAct.getId(), "Список не пустой");
-        assertTrue(listHistory.contains(subtaskAct));
-
-        inMemoryHistoryManager.remove(1);
-        listHistory = inMemoryHistoryManager.getHistory();
-        assertEquals(6,listHistory.size(), "Список не пустой");
-        assertFalse(listHistory.contains(subtaskAct));
-    }
     @Test
     void linkLast() {
         InMemoryHistoryManager inMemoryHistoryManager = new InMemoryHistoryManager();
